Finalize shop functionality.

+++ app/src/main/java/com/example/nbu/service/pojos/ITradable.java
@@ -2,6 +2,6 @@ package com.example.nbu.service.pojos;
 
 public interface ITradable {
 
-    double sellPrice();
-    double buyPrice();
+    int sellPrice();
+    int buyPrice();
 }
